pyproject.toml
[tool.poetry]
name = "testtaskgpn"
version = "0.7.0"
description = ""
authors = ["Gutes Tod <jefflol003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
sqlalchemy = "^2.0.27"
python-dotenv = "^1.0.1"
uvicorn = "^0.27.1"
gunicorn = "^21.2.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}


[tool.poetry.group.test.dependencies]
locust = "^2.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
